+++ test/TensorFlowRuntime/dynamic_compilation_tensor_group.swift
@@ -134,7 +134,7 @@ func some_tf_op(n : Int) {
   }
 
   let actual: Tensor<Float> = #tfop("Pack",
-    [arr], T$dtype: Float.tensorFlowDataType, axis: 0)
+    [arr], T$dtype: Float.tensorFlowDataType, axis: Int64(0))
   let expected  = ShapedArray<Float>(shape: [n, 3], scalars: arr_exp)
   expectEqual(expected, actual.array)
 }
